Adds actions for all verbs
