+++ bookwyrm/tests/views/test_follow.py
@@ -105,6 +105,8 @@ class BookViews(TestCase):
         request.user = self.local_user
         self.remote_user.followers.add(self.local_user)
         self.assertEqual(self.remote_user.followers.count(), 1)
+        # need to see if this ACTUALLY broadcasts
+        raise ValueError()
         with patch('bookwyrm.models.activitypub_mixin.broadcast_task.delay'):
             views.unfollow(request)
