+++ bookwyrm/views/follow.py
@@ -22,8 +22,6 @@ def follow(request):
         user_object=to_follow,
     )
 
-    if to_follow.local and not to_follow.manually_approves_followers:
-        rel.accept()
     return redirect(to_follow.local_path)
